Made the test framework able to manually shut down network. Also added documentation a few places
