+++ core/src/main/java/dk/alexandra/fresco/framework/sce/SecureComputationEngineImpl.java
@@ -43,12 +43,9 @@ import dk.alexandra.fresco.framework.network.ScapiNetworkImpl;
 import dk.alexandra.fresco.framework.sce.configuration.ProtocolSuiteConfiguration;
 import dk.alexandra.fresco.framework.sce.configuration.SCEConfiguration;
 import dk.alexandra.fresco.framework.sce.resources.ResourcePoolImpl;
-import dk.alexandra.fresco.framework.sce.resources.storage.StreamedStorage;
 import dk.alexandra.fresco.suite.ProtocolSuite;
 import java.io.IOException;
-import java.security.SecureRandom;
 import java.util.Map;
-import java.util.Random;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -88,53 +85,6 @@ public class SecureComputationEngineImpl implements SecureComputationEngine {
 
     this.evaluator = this.sceConf.getEvaluator();
     this.evaluator.setMaxBatchSize(sceConf.getMaxBatchSize());
-
-  }
-
-  private static Network getNetworkFromConfiguration(SCEConfiguration sceConf,
-      int myId, Map<Integer, Party> parties) {
-    int channelAmount = 1;
-    NetworkConfiguration conf = new NetworkConfigurationImpl(myId, parties);
-    return buildNetwork(conf, channelAmount, sceConf.getNetworkStrategy());
-  }
-
-  private static Network buildNetwork(NetworkConfiguration conf,
-      int channelAmount, NetworkingStrategy networkStrat) {
-    Network network;
-    switch (networkStrat) {
-      case KRYONET:
-        // TODO[PSN]
-        // This might work on mac?
-        network = new KryoNetNetwork();
-        //network = new ScapiNetworkImpl();
-        break;
-      case SCAPI:
-        network = new ScapiNetworkImpl();
-        break;
-      default:
-        throw new ConfigurationException("Unknown networking strategy " + networkStrat);
-    }
-    network.init(conf, channelAmount);
-    return network;
-  }
-
-  public static ResourcePoolImpl createResourcePool(SCEConfiguration sceConf) throws IOException {
-    int myId = sceConf.getMyId();
-    Map<Integer, Party> parties = sceConf.getParties();
-
-    StreamedStorage streamedStorage = sceConf.getStreamedStorage();
-    // Secure random by default.
-    Random rand = new Random(0);
-    SecureRandom secRand = new SecureRandom();
-
-    Network network = getNetworkFromConfiguration(sceConf, myId, parties);
-    network.connect(10000);
-
-    ResourcePoolImpl resourcePool =
-        new ResourcePoolImpl(myId, parties.size(),
-            network, streamedStorage, rand, secRand);
-
-    return resourcePool;
 
   }
