+++ core/src/test/java/dk/alexandra/fresco/lib/arithmetic/SearchingTests.java
@@ -32,7 +32,7 @@ import dk.alexandra.fresco.framework.TestApplication;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadConfiguration;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
-import dk.alexandra.fresco.framework.sce.SecureComputationEngineImpl;
+import dk.alexandra.fresco.framework.network.NetworkCreator;
 import dk.alexandra.fresco.framework.value.OInt;
 import dk.alexandra.fresco.framework.value.SInt;
 import dk.alexandra.fresco.lib.collections.LookUpProtocolFactory;
@@ -86,7 +86,7 @@ public class SearchingTests {
             }
           };
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
           for (int i = 0; i < PAIRS; i++) {
             final int counter = i;
             TestApplication app1 = new TestApplication() {
@@ -116,7 +116,7 @@ public class SearchingTests {
             };
 
             secureComputationEngine
-                .runApplication(app1, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+                .runApplication(app1, NetworkCreator.createResourcePool(conf.sceConf));
 
             Assert.assertEquals(values[i], app1.outputs[0].getValue()
                 .intValue());
