+++ core/src/test/java/dk/alexandra/fresco/lib/bool/BasicBooleanTests.java
@@ -32,7 +32,7 @@ import dk.alexandra.fresco.framework.TestBoolApplication;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadConfiguration;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
-import dk.alexandra.fresco.framework.sce.SecureComputationEngineImpl;
+import dk.alexandra.fresco.framework.network.NetworkCreator;
 import dk.alexandra.fresco.framework.value.OBool;
 import dk.alexandra.fresco.framework.value.SBool;
 import dk.alexandra.fresco.lib.helper.SingleProtocolProducer;
@@ -72,7 +72,7 @@ public class BasicBooleanTests {
           };
 
           secureComputationEngine.runApplication(app,
-              SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -132,7 +132,7 @@ public class BasicBooleanTests {
           };
 
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -193,7 +193,7 @@ public class BasicBooleanTests {
           };
 
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -238,7 +238,7 @@ public class BasicBooleanTests {
           };
 
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -312,7 +312,7 @@ public class BasicBooleanTests {
           };
 
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
