+++ core/src/test/java/dk/alexandra/fresco/lib/crypto/BristolCryptoTests.java
@@ -35,7 +35,7 @@ import dk.alexandra.fresco.framework.ProtocolProducer;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadConfiguration;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
-import dk.alexandra.fresco.framework.sce.SecureComputationEngineImpl;
+import dk.alexandra.fresco.framework.network.NetworkCreator;
 import dk.alexandra.fresco.framework.value.OBool;
 import dk.alexandra.fresco.framework.value.SBool;
 import dk.alexandra.fresco.lib.field.bool.BasicLogicFactory;
@@ -146,7 +146,7 @@ public class BristolCryptoTests {
           };
 
           secureComputationEngine
-              .runApplication(aesApp, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(aesApp, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -233,7 +233,7 @@ public class BristolCryptoTests {
           };
 
           secureComputationEngine
-              .runApplication(aesApp, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(aesApp, NetworkCreator.createResourcePool(conf.sceConf));
           if (!assertAsExpected) {
             return;
           }
@@ -320,7 +320,7 @@ public class BristolCryptoTests {
           };
 
           secureComputationEngine.runApplication(sha256App,
-              SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -408,7 +408,7 @@ public class BristolCryptoTests {
           };
 
           secureComputationEngine
-              .runApplication(md5App, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(md5App, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -487,7 +487,7 @@ public class BristolCryptoTests {
           };
 
           secureComputationEngine.runApplication(multApp,
-              SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
@@ -568,7 +568,7 @@ public class BristolCryptoTests {
           };
 
           secureComputationEngine
-              .runApplication(md5App, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(md5App, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
