+++ core/src/test/java/dk/alexandra/fresco/lib/database/EliminateDuplicatesTests.java
@@ -6,9 +6,9 @@ import dk.alexandra.fresco.framework.TestApplication;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadConfiguration;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
+import dk.alexandra.fresco.framework.network.NetworkCreator;
 import dk.alexandra.fresco.framework.sce.SCEFactory;
 import dk.alexandra.fresco.framework.sce.SecureComputationEngine;
-import dk.alexandra.fresco.framework.sce.SecureComputationEngineImpl;
 import dk.alexandra.fresco.framework.value.OInt;
 import dk.alexandra.fresco.framework.value.SInt;
 import dk.alexandra.fresco.lib.compare.ComparisonProtocolFactoryImpl;
@@ -106,7 +106,7 @@ class EliminateDuplicatesTests {
             }
           };
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
           secureComputationEngine.shutdownSCE();
           Assert.assertEquals(BigInteger.ZERO, app.getOutputs()[0].getValue());
           Assert.assertEquals(BigInteger.ZERO, app.getOutputs()[1].getValue());
@@ -185,7 +185,7 @@ class EliminateDuplicatesTests {
             }
           };
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
           secureComputationEngine.shutdownSCE();
           Assert.assertEquals(BigInteger.ONE, app.getOutputs()[0].getValue());
           Assert.assertEquals(BigInteger.ZERO, app.getOutputs()[1].getValue());
@@ -266,7 +266,7 @@ class EliminateDuplicatesTests {
             }
           };
           secureComputationEngine
-              .runApplication(app, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(app, NetworkCreator.createResourcePool(conf.sceConf));
           secureComputationEngine.shutdownSCE();
           Assert.assertEquals(BigInteger.ONE, app.getOutputs()[0].getValue());
           Assert.assertEquals(BigInteger.ZERO, app.getOutputs()[1].getValue());
