+++ core/src/test/java/dk/alexandra/fresco/lib/math/mult/BristolMultTests.java
@@ -34,7 +34,7 @@ import dk.alexandra.fresco.framework.ProtocolProducer;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadConfiguration;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
-import dk.alexandra.fresco.framework.sce.SecureComputationEngineImpl;
+import dk.alexandra.fresco.framework.network.NetworkCreator;
 import dk.alexandra.fresco.framework.value.OBool;
 import dk.alexandra.fresco.framework.value.SBool;
 import dk.alexandra.fresco.lib.crypto.BristolCryptoFactory;
@@ -148,7 +148,7 @@ public class BristolMultTests {
           };
 
           secureComputationEngine
-              .runApplication(md5App, SecureComputationEngineImpl.createResourcePool(conf.sceConf));
+              .runApplication(md5App, NetworkCreator.createResourcePool(conf.sceConf));
 
           if (!assertAsExpected) {
             return;
